UI: New team view v0
