+++ rcongui/example.env
@@ -7,4 +7,4 @@ REACT_APP_API_URL=http://localhost:3000/api/
 # VITE_CRCON_API_ENDPOINT: The path segment for API requests (e.g., 'api' results in '/api/*' routes)
 # VITE_CRCON_SERVER_URL: The complete URL of your CRCON server instance
 VITE_CRCON_API_ENDPOINT=/api/
-VITE_CRCON_SERVER_URL=http://192.168.4.30:8010
+VITE_CRCON_SERVER_URL=http://localhost:8010
